requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-ai-documentintelligence
requests
python-dateutil